+++ Hearthstone Deck Tracker/Hearthstone/Watchers.cs
@@ -291,6 +291,10 @@ namespace Hearthstone_Deck_Tracker.Hearthstone
 			if(Core.Game.IsTraditionalHearthstoneMatch)
 			{
 				Core.Overlay.HighlightPlayerDeckCards(args.CardId);
+				if(Core.Windows.PlayerWindow.IsVisible)
+				{
+					Core.Windows.PlayerWindow.HighlightPlayerDeckCards(state.CardId);
+				}
 			}
 		}